+++ src/client.js
@@ -1,7 +1,6 @@
 /* eslint no-console:0 */
 /* global __DEVELOPMENT__ */
 import 'babel-polyfill'
-import 'normalize.css'
 import { BrowserRouter, Route } from 'react-router-dom'
 import { getGlobalEnv } from '@twreporter/core/lib/utils/global-env'
 import App from './app'
@@ -31,7 +30,10 @@ function scrollToTopAndFirePageview() {
       // the element is rendered on the page before trying to getElementById.
       setTimeout(() => {
         const id = hash.replace('#', '')
-        const element = document.getElementById(id)
+        // since id will be encoded by browser,
+        // decode it back as a string
+        const decodedID = decodeURIComponent(id)
+        const element = document.getElementById(decodedID)
         if (element) {
           element.scrollIntoView()
         }