+++ src/components/side-bar/side-bar-factory.js
@@ -1,14 +1,9 @@
 import AboutusSideBar from './aboutus-page-side-bar'
-import ArticleSideBar from './article-side-bar.js'
 import IndexPageSideBar from './index-page-side-bar'
 import SideBarHOC from '@twreporter/react-components/lib/side-bar'
 
 function getIndexPageSideBar() {
   return SideBarHOC(IndexPageSideBar)
-}
-
-function getArticleSideBar() {
-  return SideBarHOC(ArticleSideBar)
 }
 
 function getAboutUSPageSideBar() {
@@ -17,6 +12,5 @@ function getAboutUSPageSideBar() {
 
 export default {
   getIndexPageSideBar,
-  getArticleSideBar,
   getAboutUSPageSideBar
 }